client/requirements.txt
python-dotenv
requests
pydantic>=1.10